pyproject.toml
[tool.poetry]
name = "neo_dashboard"
version = "0.1.0"
description = "NeoWS NASA Api Dashboard Backend"
authors = ["Ugly Wiz <uglywizard@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
Markdown = "^3.3.3"
django-filter = "^2.4.0"
django-cors-headers = "^3.5.0"
requests = "^2.25.0"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"


[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
black = "^20.8b1"
mock = "^4.0.3"
coverage = "^5.3.1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
